Improve PHPDoc annotations and type safety across core files
- Add variable names to @var tags in plugin constants documentation
- Cast blog_id to int in Activator::activate() for type safety
- Cast gmdate() results to int in Cron::enable_run() for gmmktime()
- Initialize Options::$options property to null with proper type hint
- Update npm dependencies: @wordpress/prettier-config, @wordpress/scripts, and terser

+++ includes/util/class-cron.php
@@ -33,7 +33,7 @@ class Cron {
 	 * @param string $recurrence Frequency.
 	 */
 	public function enable_run( $hour, $min, $recurrence ) {
-		$on = gmmktime( $hour, $min, 0, gmdate( 'm' ), gmdate( 'd' ), gmdate( 'Y' ) );
+		$on = gmmktime( $hour, $min, 0, (int) gmdate( 'm' ), (int) gmdate( 'd' ), (int) gmdate( 'Y' ) );
 
 		if ( ! wp_next_scheduled( 'acc_cron_hook' ) ) {
 			wp_schedule_event( $on, $recurrence, 'acc_cron_hook' );
